Properly initialize the component id in each component
- booter logic to update its own component id
- booter logic and mkimg logic to add ids into the initargs, and to parse them out
- component_info struct acquisition within the booter
- booter logic to update the component id within the component info structure
- NOTE: we are assuming (checked via assertions) that the component info structure is
    1. in the RW section of the data section, and
    2. has the component id initialized to 0 at compile time

+++ src/kernel/include/shared/cos_types.h
@@ -352,8 +352,8 @@ typedef unsigned long vaddr_t; /* virtual address */
 typedef unsigned int  page_index_t;
 
 typedef unsigned short int spdid_t;
-typedef unsigned short int compid_t;
-typedef unsigned short int thdid_t;
+typedef unsigned long      compid_t;
+typedef unsigned long      thdid_t;
 typedef unsigned long      invtoken_t;
 typedef int                thdclosure_index_t;
 
@@ -403,7 +403,7 @@ struct cos_stack_freelists {
 
 struct cos_component_information {
 	struct cos_stack_freelists cos_stacks;
-	long                       cos_this_spd_id;
+	unsigned long              cos_this_spd_id;
 	u32_t                      cos_tmem_relinquish[COMP_INFO_TMEM];
 	u32_t                      cos_tmem_available[COMP_INFO_TMEM];
 	vaddr_t                    cos_heap_ptr, cos_heap_limit;
